Correct Datatype for Real class in expr package
The datatype in the Real class declaration has been corrected from Int to Double. This adjustment provides an accurate datatype that fits the value it is supposed to hold, ensuring that the Real class can handle double precision floating point numbers properly. This is a crucial change as it enables the Real class to function as intended.

+++ src/main/scala/expr/Real.scala
@@ -1,4 +1,4 @@
 package cl.ravenhill.scrabble.dx
 package expr
 
-class Real(override val value: Double) extends Value[Int]
+class Real(override val value: Double) extends Value[Double]
